+++ CHANGES.rst
@@ -1,7 +1,24 @@
 
 
-1.0.4dev
+1.0.6dev
 --------
+
+ - Fix class and version check for DataContainer objects.
+
+1.0.5 (23 Jun 2020)
+-------------------
+
+ - Add median combining code
+ - Make biasframes median combine by default
+ - Implemented IFU reduction hooks
+ - KCWI reduction complete up to spec2D frames
+ - Implemented new flatfield DataContainer to separate
+   pixelflat and illumflat
+
+
+1.0.4 (27 May 2020)
+-------------------
+
  - Add a script (pypeit_flux_setup) for creating fluxing,
    coadd1d and tellfit pypeit files
  - Add telluric fitting script, pypeit_tellfit